+++ projects/ngneat/error-tailor/src/lib/form-action.directive.ts
@@ -5,12 +5,19 @@ import { shareReplay, tap } from 'rxjs/operators';
 @Directive({
   selector: 'form[errorTailor]'
 })
-export class FormSubmitDirective {
+export class FormActionDirective {
   submit$: Observable<Event> = fromEvent(this.element, 'submit').pipe(
     tap(() => {
       if (this.element.classList.contains('form-submitted') === false) {
         this.element.classList.add('form-submitted');
       }
+    }),
+    shareReplay({ refCount: true, bufferSize: 1 })
+  );
+
+  reset$: Observable<Event> = fromEvent(this.element, 'reset').pipe(
+    tap(() => {
+      this.element.classList.remove('form-submitted');
     }),
     shareReplay({ refCount: true, bufferSize: 1 })
   );